CHANGES: - Desktop, IOS and Web (Just webpage loading for now) added.

+++ composeApp/src/desktopMain/kotlin/org/example/project/presentation/expectuals/getScreenDimensions.kt
@@ -3,9 +3,9 @@ package org.example.project.presentation.expectuals
 import java.awt.Toolkit
 
 actual fun getScreenHeight(): Float {
-    return Toolkit.getDefaultToolkit().screenSize.height
+    return Toolkit.getDefaultToolkit().screenSize.height.toFloat()
 }
 
 actual fun getScreenWidth(): Float {
-    return Toolkit.getDefaultToolkit().screenSize.width
+    return Toolkit.getDefaultToolkit().screenSize.width.toFloat()
 }
